Add duplicate nav to document
  In order to show the effect of the navigation menu sliding out from below the header region, I
separated the nav element from the header. However, this navigation
effect is not present on larger screen sizes. And because I think I may
run into formatting issues based on how the effect was created and the
position of the nav in the HTML, I decided to create a second nav.